git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/7.0-patches@2127 d681b350-854c-439c-a81d-c0ce20710203

+++ src/dbc/ZDbcMySqlUtils.pas
@@ -548,7 +548,7 @@ function GetMySQLColumnInfoFromFieldHandle(PlainDriver: IZMySQLPlainDriver;
   const bUseResult:boolean): TZColumnInfo;
 var
   FieldFlags: Integer;
-  FieldLength:integer;
+  FieldLength: ULong;
 begin
   if Assigned(FieldHandle) then
   begin
